@@ -10,65 +10,67 @@
     "typings",
     "src"
   ],
+  "exports": {
+    "./css": "./dist/style.css",
+    ".": "./dist/index.mjs"
+  },
   "scripts": {
+    "build:js": "vite build",
+    "build:package": "concurrently --success first \"yarn build:js\" \"yarn emit-declarations\"",
+    "emit-declarations": "vue-tsc --project tsconfig.json --emitDeclarationOnly",
     "start": "vue-cli-service serve",
-    "lint": "vue-cli-service lint",
-    "build": "rollup --config",
-    "prepare": "yarn build",
-    "tager:update": "node scripts/tager-update.js",
-    "yalc:clean": "yalc remove --all"
+    "prepare": "yarn build:package",
+    "tager:update": "yarn upgrade --latest --scope @tager"
   },
   "dependencies": {
+    "filesize": "^9.0.11",
     "lodash.isequal": "^4.5.0",
+    "lodash.kebabcase": "^4.1.1",
     "lodash.pick": "^4.4.0",
     "path-to-regexp": "^6.1.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^15.0.0",
-    "@rollup/plugin-node-resolve": "^9.0.0",
-    "@tager/admin-dynamic-field": "^0.1.18",
-    "@tager/admin-layout": "^0.5.4",
-    "@tager/admin-services": "^0.2.26",
-    "@tager/admin-ui": "^0.8.96",
-    "@types/lodash.isequal": "^4.5.5",
-    "@types/lodash.pick": "^4.4.6",
-    "@types/node": "12.x",
-    "@typescript-eslint/eslint-plugin": "^2.26.0",
-    "@typescript-eslint/parser": "^2.26.0",
-    "@vue/cli-plugin-babel": "~4.5.6",
-    "@vue/cli-plugin-eslint": "~4.5.6",
-    "@vue/cli-plugin-router": "~4.5.6",
-    "@vue/cli-plugin-typescript": "~4.5.6",
-    "@vue/cli-service": "~4.5.6",
-    "@vue/composition-api": "^1.0.0-beta.14",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/eslint-config-typescript": "^5.1.0",
-    "babel-eslint": "^10.1.0",
-    "core-js": "^3.6.5",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-vue": "^6.2.2",
-    "prettier": "^2.1.2",
-    "rollup": "^2.27.1",
-    "rollup-plugin-typescript2": "^0.27.2",
-    "rollup-plugin-vue": "^5.1.9",
-    "sass-loader": "^10.0.2",
-    "typescript": "~3.9.10",
-    "vue": "2.6.x",
-    "vue-router": "^3.5.1",
-    "vue-template-compiler": "^2.6.11"
+    "@babel/core": "^7.17.9",
+    "@babel/eslint-parser": "^7.17.0",
+    "@rushstack/eslint-patch": "^1.1.2",
+    "@tager/admin-dynamic-field": "^1.0.0",
+    "@tager/admin-layout": "^1.0.0",
+    "@tager/admin-services": "^1.0.0",
+    "@tager/admin-ui": "^1.0.1",
+    "@types/filesize": "^5.0.0",
+    "@types/lodash.isequal": "^4.5.6",
+    "@types/lodash.pick": "^4.4.7",
+    "@types/node": "16.x",
+    "@typescript-eslint/eslint-plugin": "^5.21.0",
+    "@typescript-eslint/parser": "^5.21.0",
+    "@vitejs/plugin-vue": "^2.3.1",
+    "@vue/cli-plugin-babel": "^5.0.4",
+    "@vue/cli-plugin-typescript": "^5.0.4",
+    "@vue/cli-service": "~5.0.0",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "@vue/eslint-config-typescript": "^10.0.0",
+    "concurrently": "^7.1.0",
+    "core-js": "^3.22.3",
+    "eslint": "^8.14.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-vue": "^8.7.1",
+    "pinia": "^2.0.21",
+    "sass": "^1.51.0",
+    "sass-loader": "^12.6.0",
+    "typescript": "^4.6.3",
+    "vite": "^2.9.6",
+    "vue": "^3.2.31",
+    "vue-router": "^4.0.14",
+    "vue-tsc": "^0.34.10"
   },
   "peerDependencies": {
-    "@tager/admin-dynamic-field": ">=0.1.3",
-    "@tager/admin-layout": ">=0.4.2",
-    "@tager/admin-services": ">=0.2.12",
-    "@tager/admin-ui": ">=0.7.5",
-    "@vue/composition-api": ">=1.0.0-beta.14",
-    "vue": "2.6.x"
-  },
-  "resolutions": {
-    "prettier": "2.x"
+    "@tager/admin-layout": "^1.0.0",
+    "@tager/admin-services": "^1.0.0",
+    "@tager/admin-ui": "^1.0.1",
+    "vue": "^3.2.31",
+    "vue-router": "^4.0.14"
   },
   "publishConfig": {
     "access": "public"
