@@ -1,19 +1,14 @@
-import { onMounted, SetupContext } from '@vue/composition-api';
+import { onMounted } from 'vue';
 
 import { ResourceRef, useResource } from '@tager/admin-services';
 
 import { Category } from '../typings/model';
 import { getCategories } from '../services/requests';
 
-export function useFetchCategories({
-  context,
-}: {
-  context: SetupContext;
-}): ResourceRef<Category[]> {
+export function useFetchCategories(): ResourceRef<Category[]> {
   const [fetchCategories, resource] = useResource<Category[]>({
     fetchResource: getCategories,
     initialValue: [],
-    context,
     resourceName: 'Blog categories',
   });
 
