@@ -1,19 +1,14 @@
-import { onMounted, SetupContext } from '@vue/composition-api';
+import { onMounted } from 'vue';
 
 import { Nullable, ResourceRef, useResource } from '@tager/admin-services';
 
 import { ModuleConfig } from '../typings/model';
 import { getModuleConfig } from '../services/requests';
 
-export function useFetchModuleConfig({
-  context,
-}: {
-  context: SetupContext;
-}): ResourceRef<Nullable<ModuleConfig>> {
+export function useFetchModuleConfig(): ResourceRef<Nullable<ModuleConfig>> {
   const [fetchModuleConfig, resource] = useResource<Nullable<ModuleConfig>>({
     fetchResource: getModuleConfig,
     initialValue: null,
-    context: context,
     resourceName: 'Module configuration',
   });
 
